+++ Service/ItemService.cs
@@ -27,9 +27,25 @@ namespace MyApp_MVC.Service
         public async Task<bool> DeleteItem(int id)
         {
            //var item=await _itemRepository.Getitembyid(id);
+           var item=await _itemRepository.Getitembyid(id);
+            if(item == null)
+            {
+                throw new Exception($"Item not found with id:{id}");
+            }
            await _itemRepository.DeleteItem(id);
             return true;
         }
+
+        public async Task<ItemDto> GetItemById(int id)
+        {
+            var item = await _itemRepository.Getitembyid(id);
+            if(item == null)
+            {
+                throw new Exception($"No record found for id:{id}!");
+            }
+            return _mapper.Map<ItemDto>(item);
+        }
+
 
         public async Task<List<ItemDto>> GetItemsAsync()
         {
